s/u16/usize + s/u8/usize/

+++ src/hidden_singles.rs
@@ -6,13 +6,17 @@ pub trait HiddenSingles {
     fn handle_hidden_singles(&mut self) -> usize;
 }
 
-fn find_hidden_single<T: Iterator<Item = u8>>(sudoku: &Sudoku, i: usize, iter: T) -> Option<u16> {
+fn find_hidden_single<T: Iterator<Item = usize>>(
+    sudoku: &Sudoku,
+    i: usize,
+    iter: T,
+) -> Option<usize> {
     let set = iter
         .filter_map(|idx| {
-            if idx == i as u8 {
+            if idx == i {
                 None
             } else {
-                Some(sudoku.positions[idx as usize].get_candidates())
+                Some(sudoku.positions[idx].get_candidates())
             }
         })
         .fold(0, |acc, s| acc | s);
@@ -24,15 +28,15 @@ fn find_hidden_single<T: Iterator<Item = u8>>(sudoku: &Sudoku, i: usize, iter: T
     }
 }
 
-fn find_hidden(sudoku: &Sudoku) -> Vec<(u8, u16)> {
+fn find_hidden(sudoku: &Sudoku) -> Vec<(usize, usize)> {
     (0..sudoku.positions.len())
-        .filter(|i| sudoku.positions[*i as usize].num_candidates() != 1)
+        .filter(|i| sudoku.positions[*i].num_candidates() != 1)
         .filter_map(|i| {
-            let (row_i, col_i, box_i) = get_index_tuple(i as u8);
+            let (row_i, col_i, box_i) = get_index_tuple(i);
             find_hidden_single(sudoku, i, row_iter(row_i))
                 .or_else(|| find_hidden_single(sudoku, i, col_iter(col_i)))
                 .or_else(|| find_hidden_single(sudoku, i, box_iter(box_i)))
-                .map(|h| (i as u8, h as u16))
+                .map(|h| (i, h))
         })
         .collect()
 }
@@ -45,8 +49,8 @@ impl HiddenSingles for Sudoku {
         for (i, h) in hidden {
             found += 1;
             assert_eq!(1, h.count_ones());
-            self.positions[i as usize].candidates = h;
-            self.positions[i as usize].set_solved();
+            self.positions[i].candidates = h;
+            self.positions[i].set_solved();
         }
         found
     }
